+++ Source/Database/Database.cpp
@@ -5,6 +5,9 @@ NS_ENIGMA_BEGIN
 
 void Database::Initialize()
 {
+#ifdef ENIGMA_DEBUG
+	ENIGMA_TRACE(ENIGMA_CURRENT_FUNCTION);
+#endif
 	ENIGMA_INFO("SQLite3 version {0}", SQLite::VERSION);
 	try
 	{
@@ -20,20 +23,23 @@ void Database::Initialize()
 		}
 
 		m_database = std::make_unique<SQLite::Database>(
-			Constants::Database::DATABASE_FILE_PATH,
+			Constants::Database::DATABASE_FILE_NAME,
 			SQLite::OPEN_CREATE | SQLite::OPEN_READWRITE // create if not exists
 			);
 
 		// Create Tables If Not Exists
-		const auto query = std::make_unique<SQLite::Statement>(
-			*m_database,
-			Constants::Database::CREATE_TABLES_QUERY
-			);
-		const i32 status = query->exec(); // status 0 == SQLite::OK
-		if (status != SQLite::OK)
+		for (const auto& create_table_sql : Constants::Database::CREATE_TABLES_QUERIES)
 		{
-			ENIGMA_CRITICAL("Database failed to create tables with error code: {0} | msg: {1}", query->getErrorCode(), query->getErrorMsg());
-			std::exit(EXIT_FAILURE);
+			//ENIGMA_LOG("SQL: {0}", create_table_sql);
+			const auto query = std::make_unique<SQLite::Statement>(
+				*m_database,
+				create_table_sql
+				);
+			const i32 status = query->exec(); // status 0 == SQLite::OK
+			ENIGMA_ASSERT_OR_THROW(status == SQLite::OK,
+				"Database failed to create tables with error code " +
+				std::to_string(query->getErrorCode()) + 
+				" and message: " + String(query->getErrorMsg()));
 		}
 	}
 	catch (const SQLite::Exception& e)
@@ -44,19 +50,47 @@ void Database::Initialize()
 
 }
 
+// Add Encryption to Encryptions table, returns true on success
 bool Database::AddEncryption(const std::unique_ptr<Encryption>& e)
 {
+#ifdef ENIGMA_DEBUG
+	ENIGMA_TRACE(ENIGMA_CURRENT_FUNCTION);
+#endif
+	ENIGMA_ASSERT_OR_RETURN(m_database, "Database was not initialized", false);
 	try
 	{
-		constexpr char* sql = "INSERT INTO Encryptions(title, cipher, date_time, is_file) VALUES(?, ?, DATETIME(), ?)";
-		ENIGMA_LOG("SQL: {0}", sql);
+		auto transaction = std::make_unique<SQLite::Transaction>(*m_database);
 
-		auto query = std::make_unique<SQLite::Statement>(*m_database, sql);
-		query->bindNoCopy(1, e->title);
-		query->bindNoCopy(2, e->cipher);
-		query->bind(3, static_cast<i32>(e->is_file));
-		i32 r = query->exec(); // returns # of rows effected
-		return r > 0;
+		// Insert encryption
+		{
+			constexpr char* sql = "INSERT INTO Encryption(title, date_time, is_file) VALUES(?, DATETIME(), ?)";
+			ENIGMA_LOG("SQL: {0}", sql);
+			auto query = std::make_unique<SQLite::Statement>(*m_database, sql);
+			query->bindNoCopy(1, e->title);
+			query->bind(2, static_cast<i32>(e->is_file));
+			i32 r = query->exec(); // returns # of rows effected
+			ENIGMA_ASSERT_OR_THROW(r > 0, "Failed to insert encyption record");
+		}
+		
+		// Get inserted encryption id
+		const i64 encryption_last_inserted_id = m_database->getLastInsertRowid();
+		
+		// Insert cipher
+		{
+			constexpr char* sql = "INSERT INTO Cipher(data, ide) VALUES(?, ?)";
+			ENIGMA_LOG("SQL: {0}", sql);
+			auto query = std::make_unique<SQLite::Statement>(*m_database, sql);
+			query->bindNoCopy(1, e->cipher.data.data(), e->cipher.data.size()); // bind blob
+			query->bind(2, encryption_last_inserted_id);
+			i32 r = query->exec(); // returns # of rows effected
+			ENIGMA_ASSERT_OR_THROW(r > 0, "Failed to insert cipher record");
+		}
+		
+		
+		// OK
+		transaction->commit();
+
+		return true;
 	}
 	catch (const SQLite::Exception& e)
 	{
@@ -66,6 +100,48 @@ bool Database::AddEncryption(const std::unique_ptr<Encryption>& e)
 
 }
 
+// Delete Encryption record by id, returns true if successfully deleted
+bool Database::DeleteEncryption(const i64 ide)
+{
+#ifdef ENIGMA_DEBUG
+	ENIGMA_TRACE(ENIGMA_CURRENT_FUNCTION);
+#endif
+	ENIGMA_ASSERT_OR_RETURN(m_database, "Database was not initialized", false);
+	try
+	{
+		auto transaction = std::make_unique<SQLite::Transaction>(*m_database);
+
+		// Delete cipher
+		{
+			constexpr char* sql = "DELETE FROM Cipher WHERE ide = ?";
+			ENIGMA_LOG("SQL: {0}", sql);
+			auto query = std::make_unique<SQLite::Statement>(*m_database, sql);
+			query->bind(1, ide);
+			i32 r = query->exec(); // returns # of rows effected
+			ENIGMA_ASSERT_OR_THROW(r > 0, "Failed to delete cipher record");
+		}
+
+		// Delete encryption
+		{
+			constexpr char* sql = "DELETE FROM Encryption WHERE ide = ?";
+			ENIGMA_LOG("SQL: {0}", sql);
+			auto query = std::make_unique<SQLite::Statement>(*m_database, sql);
+			query->bind(1, ide);
+			i32 r = query->exec(); // returns # of rows effected
+			ENIGMA_ASSERT_OR_THROW(r > 0, "Failed to delete encyption record");
+		}
+
+		// OK
+		transaction->commit();
+
+		return true;
+	}
+	catch (const SQLite::Exception& e)
+	{
+		ENIGMA_ERROR(e.what());
+		return false;
+	}
+}