+++ Source/Scenes/MyEncryptionsScene.hpp
@@ -13,16 +13,17 @@ public:	/* Constructors / Destructor */
 
 private: /* Overrides */
 	void OnCreate() override;
-	[[maybe_unused]] void OnUpdate(const f32 & dt) override;
+	[[maybe_unused]] void OnUpdate(const f32&) override;
 	void OnDraw() override;
 	void OnImGuiDraw() override;
-	void OnEvent(Enigma::Event & event) override;
+	[[maybe_unused]] void OnEvent(Event&) override;
 	void OnDestroy() override;
 
 private: /* Callbacks */
 	void OnBackButtonPressed();
-	void OnViewEncryptionButtonPressed(const size_t id);
-	void OnDeleteEncryptionButtonPressed(const size_t id);
+	void OnViewEncryptionButtonPressed(const size_t ide);
+	// returns true if item deleted successfully to notify draw loop that vector range changed
+	bool OnDeleteEncryptionButtonPressed(const size_t ide); 
 
 private: /* Initializers */
 	// Get all Encryptions from database
@@ -33,19 +34,10 @@ private:
 
 private:
 	std::vector<std::unique_ptr<Encryption>> m_encryptions;
-	
-	enum class OrderBy : byte
-	{
-		ID,
-		Title,
-		DateTime
-	} m_order_by = OrderBy::ID;
 
-	enum class Order : byte
-	{
-		Ascending,
-		Descending,
-	} m_order = Order::Descending;
+	Database::OrderBy m_order_by = Database::OrderBy::ID;
+	Database::Order m_order = Database::Order::Descending;
+
 };
 
 NS_ENIGMA_END