+++ CHANGELOG.md
@@ -25,4 +25,5 @@ If such a setting is enabled, glfwSwapInterval will have no effect." said GLFW.
 - Fixed Clang compilation issue on linux requires pic flag '-fPIC' for libraries (inih, curl, zlib, mbedlts)
 - Added CLI option  (--updates | -n)  which checks for updates when using Command line interface
 - Upgraded ImGui from v1.79 to v1.81 master branch releases https://github.com/ocornut/imgui/releases/tag/v1.81
+- Added Database system to save encryption records
 -->